fix pixel addressing on 32-bit platforms
there was a size_t where there should have been a guint64, causing
addressing problems on very large images on 32-bit platforms (like
win32)

looks like this had been broken since March 2011

+++ libvips/include/vips/image.h
@@ -407,7 +407,7 @@ typedef struct _VipsImageClass {
 
 GType vips_image_get_type( void );
 
-extern const size_t vips__image_sizeof_bandformat[];
+extern const guint64 vips__image_sizeof_bandformat[];
 
 /* Pixel address calculation macros.
  */
